CodingAgain/perplexity-gemini-clone/backend/requirements.txt
# backend/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai